Address sleep mode backend review comments
