+++ tests/v1/worker/test_gpu_worker.py
@@ -6,6 +6,11 @@ from types import SimpleNamespace
 import pytest
 
 import vllm.v1.worker.gpu_worker as gpu_worker_module
+from vllm.device_allocator.sleep_mode_backend import (
+    CuMemBackend,
+    SleepModeBackend,
+    SleepModeBackendFactory,
+)
 from vllm.multimodal.video import (
     PYNVVIDEOCODEC_CUDA_CONTEXT_BYTES,
     PYNVVIDEOCODEC_DECODER_GPU_MEMORY_BYTES,
@@ -44,6 +49,67 @@ def _pynvvideocodec_decoder_budget(api_process_count: int = 1) -> int:
         PYNVVIDEOCODEC_DECODER_GPU_MEMORY_BYTES * PYNVVIDEOCODEC_MAX_RETAINED_DECODERS
         + PYNVVIDEOCODEC_CUDA_CONTEXT_BYTES
     )
+
+
+def test_sleep_mode_cumem_is_the_default_registered_backend():
+    backend_cls = SleepModeBackendFactory.get_backend_class("cumem")
+    assert backend_cls is CuMemBackend
+    assert issubclass(backend_cls, SleepModeBackend)
+    assert backend_cls.is_supported() is True
+
+
+def test_sleep_mode_new_backend_starts_in_running_state():
+    # Constructing a backend must not touch the GPU; only suspend/resume do.
+    assert CuMemBackend().state() == "RUNNING"
+
+
+def test_sleep_mode_unknown_backend_raises():
+    with pytest.raises(ValueError, match="Unsupported sleep-mode backend"):
+        SleepModeBackendFactory.get_backend_class("does-not-exist")
+
+
+def test_sleep_mode_duplicate_registration_raises():
+    with pytest.raises(ValueError, match="already registered"):
+        SleepModeBackendFactory.register_backend(
+            "cumem",
+            "vllm.device_allocator.sleep_mode_backend",
+            "CuMemBackend",
+        )
+
+
+def test_sleep_mode_third_party_backend_registration_and_resolution():
+    """A plugin registers a backend by name; the factory resolves it lazily."""
+    name = "_pytest_dummy_backend"
+    try:
+        SleepModeBackendFactory.register_backend(
+            name,
+            __name__,
+            "DummySleepModeBackend",
+        )
+        resolved = SleepModeBackendFactory.get_backend_class(name)
+        assert resolved is DummySleepModeBackend
+    finally:
+        SleepModeBackendFactory._registry.pop(name, None)
+
+
+def test_sleep_mode_suspend_resume_state_transitions():
+    """Lifecycle state advances RUNNING -> SUSPENDED -> RUNNING without GPU."""
+    backend = DummySleepModeBackend()
+    assert backend.state() == "RUNNING"
+    backend.suspend(level=1)
+    assert backend.state() == "SUSPENDED"
+    backend.resume()
+    assert backend.state() == "RUNNING"
+
+
+class DummySleepModeBackend(SleepModeBackend):
+    """A no-GPU backend used to exercise lifecycle + registration in CPU tests."""
+
+    def suspend(self, level: int = 1) -> None:
+        self._state = "SUSPENDED"
+
+    def resume(self, tags: list[str] | None = None) -> None:
+        self._state = "RUNNING"
 
 
 @pytest.mark.parametrize("video_backend", [None, "opencv"])
